Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/084-expected.txt

CONSOLE MESSAGE: line 1: TypeError: null is not an object (evaluating 'top.log')
CONSOLE MESSAGE: line 1515: Error: Tried to create a test with file_is_test
FAILED (This TC requires JavaScript enabled)

FAIL  scheduler: event listener defined by script in a removed IFRAME TypeError: null is not an object (evaluating 'top.log')
